Added feature to change the luminosity function based on RA, Dec
For some use cases, e.g. magnification of SNe by gravitational lenses,
the luminosity function of the transient needs to be aware of the
coordinates of the transient. Previously this information had not been
passed to it. This commit add RA and Dec as keyword arguments to be
passed to the function drawing the LC parameters from the luminosity
function etc. Because of this the function definition must now either
include `ra` and `dec` as arguments or use `**kwargs` to catch them.