Added methods to remove and add transients lightcurve parameters
In order to manitpulate transient lightcurve parameter, e.g. for
making ligthcurves of multiple images when simulationg strong lensing,
method to remove and add lightcurve parameters are needed. In
combination with get_lc_param() these functions will allow to get the
parameters of a given lightcurve, remove that transient, manipulate
the parameters, e.g. according to time delay and magnification, and
then add the new version(s) of that transient.