Added function for setting individual parameters and limited get function to individual transients
This commit adds a function that allows individual parameters to be
set to new values. Additional the functions introduced in [0] were
changed to only allow a single index. This was necessary because the
selection of multiple indices did not work as intended. As a
consequence get_lightcurve_full_param() is restored and no longer
deprecated, reverting [1].

[0] - f00ce5c
[1] - d714288